Data Analysis & Visualization
The Italian scenario of solid waste: composition, recycling rates and segregation. This report looked in depth into the waste production and recycling of Italy, one of the countries of the European Union with the highest rate of waste sorting. The dataset was first explored through Descriptive and Inferential Statistic to identify the least performant areas of Italy, and relationship between the production and recycling rates; then, through Machine Learning forecasting algorithms, the values are predicted for the next years to understand the future amount of waste produced by the country. As one the first step for a better recycling starts with waste sorting in citizens everyday life, Deep Learning lightweight classification neural networks (MobileNet and DenseNet121) are applied to a dataset of trash images in the context of Smart Bins for Automated Waste Sorting. Evidence-based recommendations are given at the end of the analysis.