Eco-Engineering
This report aims to design a 100% recyclable and 100% made of recycled materials reusable cup of high-quality materials from a controlled supply chain. As a matter of fact people’s awareness is often opposite to their shaky trust in the process that brought disposed plastic into the product they are holding in their hands.
The 3rCup not only is a eco-friendly cup that cares about the environment, but also a responsible company that takes care of the raw materials from the moment they are inputted into the supply chain until their end-of-life – which is not represented by the end of life of their product – but the end of life of the material as it is not recyclable anymore.