Intelligent Systems and Robotics
Autonomous delivery robots. In the era of e-commerce and food delivery the traditional supply chain has been profoundly modified. Not only is it now fundamental to keep track of every stage of the process, but also is important to plan the chain from the warehouse down to the customer’s doorstep with the so-called last-mile phase.
From the challenges that this difficult phase has always raised, many solutions have been developed and deployed as alternatives to the traditional deliveries through couriers, postmen, and riders. With COVID-19 pandemic, the evolution of the autonomous systems took a leap forward, as it became the perfect mean to avoid social interaction. Along with the always current sustainability issue, this paper analyses the structure of the state-of- the art delivery robots, its limits, its possible improvements and future works.